[21] extract word search & sort model part
